% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{bidask-package}
\alias{bidask}
\alias{bidask-package}
\title{bidask: Efficient Estimation of Bid-Ask Spreads from Open, High, Low, and Close Prices}
\description{
Implements an efficient estimator of bid-ask spreads from open, high, low, and close prices as described in Ardia, Guidotti, & Kroencke (2021) \url{https://www.ssrn.com/abstract=3892335}. It also provides an implementation of the estimators described in Roll (1984) \doi{10.1111/j.1540-6261.1984.tb03897.x}, Corwin & Schultz (2012) \doi{10.1111/j.1540-6261.2012.01729.x}, and Abdi & Ranaldo (2017) \doi{10.1093/rfs/hhx084}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eguidotti/bidask}
  \item Report bugs at \url{https://github.com/eguidotti/bidask/issues}
}

}
\author{
\strong{Maintainer}: Emanuele Guidotti \email{emanuele.guidotti@unine.ch} (\href{https://orcid.org/0000-0002-8961-6623}{ORCID})

Other contributors:
\itemize{
  \item David Ardia (\href{https://orcid.org/0000-0003-2823-782X}{ORCID}) [contributor]
  \item Tim Kroencke (\href{https://orcid.org/0000-0001-8700-356X}{ORCID}) [contributor]
}

}
\keyword{internal}
