% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.R
\name{CS}
\alias{CS}
\title{Corwin-Schultz Estimator}
\usage{
CS(x, width = nrow(x), method = "CS", signed = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{\code{xts} object with columns named \code{High}, \code{Low}, \code{Close}, representing HLC prices.}

\item{width}{integer width of the rolling window to use, or vector of endpoints defining the intervals to use.}

\item{method}{one of \code{"CS"}, \code{"CS2"}.}

\item{signed}{a \code{logical} value indicating whether signed estimates should be returned.}

\item{na.rm}{a \code{logical} value indicating whether \code{NA} values should be stripped before the computation proceeds. Default \code{FALSE}.}
}
\value{
Time series of spread estimates.
}
\description{
Corwin-Schultz Estimator
}
\references{
Corwin, S. A., & Schultz, P. (2012). A simple way to estimate bid-ask spreads from daily high and low prices. The Journal of Finance, 67 (2), 719-760.
\doi{10.1111/j.1540-6261.2012.01729.x}
}
\keyword{internal}
