% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{EDGE}
\alias{EDGE}
\title{EDGE Estimator}
\usage{
EDGE(x, width = nrow(x), probs = c(0.025, 0.975), na.rm = FALSE, trim = 0)
}
\arguments{
\item{x}{\code{xts} object with columns named \code{Open}, \code{High}, \code{Low}, \code{Close}, representing OHLC prices.}

\item{width}{integer width of the rolling window to use, or vector of endpoints defining the intervals to use.}

\item{probs}{vector of probabilities to compute the critical values.}

\item{na.rm}{a \code{logical} value indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each end before the spread is computed. Values of trim outside that range are taken as the nearest endpoint.}
}
\value{
Time series of spread estimates.
}
\description{
EDGE Estimator
}
\keyword{internal}
