% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{edge_expanding}
\alias{edge_expanding}
\title{Expanding Estimates of Bid-Ask Spreads from Open, High, Low, and Close Prices}
\usage{
edge_expanding(open, high, low, close, sign = FALSE, na.rm = TRUE)
}
\arguments{
\item{open}{numeric vector of open prices.}

\item{high}{numeric vector of high prices.}

\item{low}{numeric vector of low prices.}

\item{close}{numeric vector of close prices.}

\item{sign}{whether to return signed estimates.}

\item{na.rm}{whether to ignore missing values.}
}
\value{
Vector of spread estimates. 
A value of 0.01 corresponds to a spread of 1\%.
This function always returns a result of the same length as the input prices.
}
\description{
Implements an expanding window calculation of the efficient estimator of bid-ask spreads 
from open, high, low, and close prices described in Ardia, Guidotti, & Kroencke (JFE, 2024):
\doi{10.1016/j.jfineco.2024.103916}.
}
\details{
Prices must be sorted in ascending order of the timestamp.
}
\examples{
# simulate open, high, low, and close prices with spread 1\%
x <- sim(n = 1000, spread = 0.01)

# estimate the spread using an expanding window
s <- edge_expanding(x$Open, x$High, x$Low, x$Close)
tail(s)
# equivalent to
s <- edge_rolling(x$Open, x$High, x$Low, x$Close, width = 1:nrow(x), na.rm = TRUE)
tail(s)

}
\references{
Ardia, D., Guidotti, E., Kroencke, T.A. (2024). Efficient Estimation of Bid-Ask Spreads from Open, High, Low, and Close Prices. Journal of Financial Economics, 161, 103916. 
\doi{10.1016/j.jfineco.2024.103916}
}
