% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{GMM}
\alias{GMM}
\title{GMM Estimator}
\usage{
GMM(x, width = nrow(x), na.rm = FALSE)
}
\arguments{
\item{x}{\code{xts} object with columns named \code{Open}, \code{High}, \code{Low}, \code{Close}, representing OHLC prices.}

\item{width}{integer width of the rolling window to use, or vector of endpoints defining the intervals to use.}

\item{na.rm}{a \code{logical} value indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
Time series of spread estimates.
}
\description{
GMM Estimator
}
\keyword{internal}
